renamed peak fitting to segmentation imported two additional sub-functions: targ_rec and simple_connectivity
