+++ liboptv/include/segmentation.h
@@ -7,8 +7,14 @@
         pp.165-168
 */
 
-#ifndef PEAKFITTING_H
-#define PEAKFITTING_H
+#ifndef SEGMENTATION_H
+#define SEGMENTATION_H
+
+#include "tracking_frame_buf.h"
+#include "parameters.h" 
+#include <stdlib.h>
+#include <stdio.h>
+
 
 typedef struct
 {
@@ -28,13 +34,20 @@ typedef struct
 }
 targpix;
 
-#include "tracking_frame_buf.h"
-#include "parameters.h" 
-#include <stdlib.h>
-#include <stdio.h>
 
-int peak_fit_new ( unsigned char *img, char par_file[], int xmin, int xmax, int ymin, 
+int peak_fit_new ( unsigned char *img, target_par *trgtpar, int xmin, int xmax, int ymin, 
 int ymax, target pix[], int nr, control_par *cpar); 
+
+void check_touch (peak *tpeak, int p1, int p2);
+
+void simple_connectivity(unsigned char *img0,
+    unsigned char *img, char par_file[],
+    int xmin, int xmax, int ymin, int ymax,
+    target pix[], int nr, int *num, control_par *cpar);
+
+void targ_rec (unsigned char *img0, unsigned char *img, char par_file[], 
+    int xmin, int xmax, int ymin, int ymax,
+    target pix[], int nr, int *num, control_par *cpar);
 
 #endif
